Fix output of pressure stall time with cgroups v2
Because of the use of type float,
for small values (which can occur here)
the output would be in scientific notation,
but we promise decimal notation.
Furthermore, float can of course be imprecise.
So we switch to the type Decimal and make sure it is printed properly.

It would be good to also switch to Decimal for other time values,
but we should make that change separately.

+++ benchexec/cgroupsv2.py
@@ -13,7 +13,7 @@ import signal
 import tempfile
 import threading
 import time
-
+from decimal import Decimal
 
 from benchexec import util
 from benchexec.cgroups import Cgroups
@@ -334,6 +334,7 @@ class CgroupsV2(Cgroups):
         """
         cpu_stats = dict(self.get_key_value_pairs(self.CPU, "stat"))
 
+        # TODO switch to Decimal together with all other float values
         return float(cpu_stats["usage_usec"]) / 1_000_000
 
     def read_max_mem_usage(self):
@@ -346,21 +347,21 @@ class CgroupsV2(Cgroups):
         mem_some_stats = mem_stats["some"].split(" ")
         stats_map = {s[0]: s[1] for s in (s.split("=") for s in mem_some_stats)}
 
-        return float(stats_map["total"]) / 1_000_000
+        return Decimal(stats_map["total"]) / 1_000_000
 
     def read_cpu_pressure(self):
         cpu_stats = dict(self.get_key_value_pairs(self.CPU, "pressure"))
         cpu_some_stats = cpu_stats["some"].split(" ")
         stats_map = {s[0]: s[1] for s in (s.split("=") for s in cpu_some_stats)}
 
-        return float(stats_map["total"]) / 1_000_000
+        return Decimal(stats_map["total"]) / 1_000_000
 
     def read_io_pressure(self):
         io_stats = dict(self.get_key_value_pairs(self.IO, "pressure"))
         io_some_stats = io_stats["some"].split(" ")
         stats_map = {s[0]: s[1] for s in (s.split("=") for s in io_some_stats)}
 
-        return float(stats_map["total"]) / 1_000_000
+        return Decimal(stats_map["total"]) / 1_000_000
 
     def read_usage_per_cpu(self):
         logging.debug("Usage per CPU not supported in cgroups v2")
